requirements.txt
asgiref==3.3.1
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
cryptography==35.0.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.1.4
django-braces==1.14.0
django-cors-headers==3.6.0
django-dotenv==1.4.2
django-oauth-toolkit==1.5.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.8.0
drf-social-oauth2==1.0.9
geographiclib==1.52
geopy==2.1.0
gunicorn==20.0.4
idna==2.10
jwcrypto==1.0
oauthlib==3.1.1
Pillow==8.4.0
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.2.0
python3-openid==3.2.0
pytz==2020.5
qrcode==7.3.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.7
whitenoise==5.3.0
wrapt==1.13.1
